refactor: split interface/impl files; remove platform attr from interface
- Move each concrete adapter class to its own file (CvarRegistry.cs,
  CvarFileService.cs, GameWindowService.cs) to match the project's
  one-class-per-file convention
- Rename DefaultCvarRegistry → CvarRegistry (no other concrete class
  uses the Default prefix)
- Change ICvarRegistry return types from T[] to IReadOnlyList<T>
- Remove [SupportedOSPlatform("windows")] from IGameWindowService —
  the attribute belongs only on the concrete P/Invoke class so that
  implementers (stubs, tests) are not forced to carry a platform
  annotation
