+++ App.axaml.cs
@@ -71,7 +71,7 @@ public partial class App : Application
             };
 
             var services = new ServiceCollection();
-            services.AddSingleton<ICvarRegistry, DefaultCvarRegistry>();
+            services.AddSingleton<ICvarRegistry, CvarRegistry>();
             services.AddSingleton<ICvarFileService, CvarFileService>();
             services.AddSingleton<IGameWindowService, GameWindowService>();
             services.AddSingleton<ISteamManager, SteamManager>();
