+++ Preview/PreviewStubs.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Collections.Generic;
-using System.Runtime.Versioning;
 using System.Threading;
 using System.Threading.Tasks;
 using d2c_launcher.Api;
@@ -298,7 +297,6 @@ internal sealed class StubTriviaRepository : ITriviaRepository
         ]);
 }
 
-[SupportedOSPlatform("windows")]
 internal sealed class StubGameWindowService : IGameWindowService
 {
     public bool IsWindowOpen() => false;
